Reorganize code to ensure no circular dependencies and no import of pytest_wdl.core except by pytest_wdl.fixtures; make url schemes pluggable and add dx:// implementation
