+++ .coveragerc
@@ -14,3 +14,4 @@ exclude_lines =
     raise NotImplementedError
     if __name__ == .__main__.:
     LOG.*
+    pass
