+++ pytest_wdl/executors/__init__.py
@@ -10,14 +10,72 @@
 #    distributed under the License is distributed on an "AS IS" BASIS,
 #    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #    See the License for the specific language governing permissions and
-#    limitations under the License.from abc import ABCMeta
+#    limitations under the License.
+from abc import ABCMeta, abstractmethod
 import json
 from pathlib import Path
 import tempfile
-from typing import Optional, Sequence, Tuple
+from typing import List, Optional, Sequence, Tuple
 
-from pytest_wdl.core import DataFile
-from pytest_wdl.utils import ensure_path
+from pytest_wdl.data_types import DataFile
+from pytest_wdl.utils import ensure_path, safe_string
+
+from WDL import Tree
+
+
+class Executor(metaclass=ABCMeta):
+    """
+    Base class for WDL workflow executors.
+
+    Args:
+        import_dirs: Relative or absolute paths to directories containing WDL
+            scripts that should be available as imports.
+    """
+    def __init__(self, import_dirs: Optional[List[Path]] = None):
+        self.import_dirs = import_dirs or []
+
+    def _get_workflow_name(self, wdl_path: Path, kwargs: dict):
+        if "workflow_name" in kwargs:
+            return kwargs["workflow_name"]
+        elif Tree:
+            if "check_quant" not in kwargs:
+                kwargs["check_quant"] = False
+            doc = Tree.load(
+                str(wdl_path),
+                path=[str(path) for path in self.import_dirs],
+                **kwargs
+            )
+            return doc.workflow.name
+        else:
+            return safe_string(wdl_path.stem)
+
+    @abstractmethod
+    def run_workflow(
+        self,
+        wdl_path: Path,
+        inputs: Optional[dict] = None,
+        expected: Optional[dict] = None,
+        **kwargs
+    ) -> dict:
+        """
+        Run a WDL workflow on given inputs, and check that the output matches
+        given expected values.
+
+        Args:
+            wdl_path: The WDL script to execute.
+            inputs: Object that will be serialized to JSON and provided to Cromwell
+                as the workflow inputs.
+            expected: Dict mapping output parameter names to expected values.
+            kwargs: Additional executor-specific keyword arguments (mostly for
+                debugging)
+
+        Returns:
+            Dict of outputs.
+
+        Raises:
+            Exception: if there was an error executing the workflow
+            AssertionError: if the actual outputs don't match the expected outputs
+        """
 
 
 def get_workflow_inputs(
